transfer user interface to show search menu to new class
